Remove rowmappers for level, criterion, question, link; changed queries; rowmappers moved to dao/extractor, MaterialConvertor renamed to MaterialMapper
