+++ src/main/java/com/softserve/betterlearningroom/dao/QuestionDao.java
@@ -1,11 +1,11 @@
 package com.softserve.betterlearningroom.dao;
 
-import com.softserve.betterlearningroom.dao.extractor.QuestionRowMapper;
 import com.softserve.betterlearningroom.entity.Material;
 import com.softserve.betterlearningroom.entity.Question;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -32,7 +32,7 @@ public class QuestionDao {
     private String removeQuery;
 
     public List<Question> getAllQuestions(Long materialId) {
-        return jdbcTemplate.query(getAllQuery, new MapSqlParameterSource("materialid", materialId), new QuestionRowMapper());
+        return jdbcTemplate.query(getAllQuery, new MapSqlParameterSource("materialid", materialId), BeanPropertyRowMapper.newInstance(Question.class));
     }
 
     public List<Question> getAllQuestions(Material material) {
