+++ src/main/java/com/softserve/betterlearningroom/dao/extractor/UserRowMapper.java
@@ -1,0 +1,26 @@
+package com.softserve.betterlearningroom.mapper;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Component;
+
+import com.softserve.betterlearningroom.entity.User;
+
+@Component
+public class UserRowMapper implements RowMapper<User> {
+
+	@Override
+	public User mapRow(ResultSet rs, int rowNum) throws SQLException {
+		User user = new User();
+		user.setId(rs.getInt("id"));
+		user.setFirstName(rs.getString("firstname"));
+		user.setLastName(rs.getString("lastname"));
+		user.setEmail(rs.getString("email"));
+		user.setPassword(rs.getString("password"));
+		user.setEnabled(rs.getBoolean("enabled"));
+		return user;
+	}
+
+}
