+++ src/main/resources/materialQuery.properties
@@ -1,27 +1,19 @@
 get.all = SELECT mtr.materialid, mtr.materialtext, mtr.startdate,\
   mtr.duedate, mtr.task, mtr.materialType, mtr.maxScore,\
-  mtr.testUrl, mtr.title, cr.criterionid, cr.title, cr.description,\
-  lv.levelid, lv.title as leveltitle, lv.description as leveldescription,\
+  mtr.testUrl, mtr.title, mtr.topicid,\
   lin.linkid, lin.url, lin.linktext, q.questionid, q.question\
   FROM materials as mtr\
   FULL JOIN links as lin ON mtr.materialid=lin.materialid\
   FULL JOIN questions as q ON mtr.materialid=q.materialid\
-  FULL JOIN criterions as cr ON mtr.materialid=cr.materialid\
-  FULL JOIN levels as lv ON cr.criterionid=lv.criterionid\
-  WHERE q.materialid=:materialid OR lin.materialid=:materialid\
-  OR cr.materialid=:materialid AND lv.criterionid=cr.criterionid
+  WHERE mtr.topicid=:topicid
 get.by.id = SELECT mtr.materialid, mtr.materialtext, mtr.startdate,\
   mtr.duedate, mtr.task, mtr.materialType, mtr.maxScore,\
-  mtr.testUrl, mtr.title, cr.criterionid, cr.title, cr.description,\
-  lv.levelid, lv.title as leveltitle, lv.description as leveldescription,\
-  lin.linkid, lin.url, lin.linktext, q.questionid, q.question\
+  mtr.testUrl, mtr.title, lin.linkid, lin.url, lin.linktext,\
+  q.questionid, q.question\
   FROM materials as mtr\
   FULL JOIN links as lin ON mtr.materialid=lin.materialid\
   FULL JOIN questions as q ON mtr.materialid=q.materialid\
-  FULL JOIN criterions as cr ON mtr.materialid=cr.materialid\
-  FULL JOIN levels as lv ON cr.criterionid=lv.criterionid\
-  WHERE q.materialid=:materialid OR lin.materialid=:materialid\
-  OR cr.materialid=:materialid AND lv.criterionid=cr.criterionid
+  WHERE q.materialid=:materialid OR lin.materialid=:materialid
 add.new = INSERT INTO materials (title, materialtext, startdate, duedate, task, materialtype, maxscore, testurl) \
   VALUES (:title, :materialtext, :startdate, :duedate, :task, :materialType, :maxScore, :testUrl)
 update = UPDATE materials SET title=:title, materialtext=:materialtext, startdate=:startdate, duedate=:duedate, \
